Don't specify URL scheme for font in flatly
By specifying a scheme you can have problems, like when a page is
running under https and you fetch the font using http - browsers
will not be happy with that. By not specifiying a scheme you the
browser will use the one it is currently using (https if https,
http if http).